shared library ============== - Change Energy historizer from double to unsigned long.
Plugin RFXCOM 
=============
- Change some conversion due to the Energy type change.

Plugin RFXLANXPL
================
- Change some conversion due to the Energy type change.

Plugin TeleInfoUSB (In Progress)
==================
- Add the first release of the teleinfo plugin.
- Work with BASE contract.
- Should work for HEURE CREUSE / HEURE PLEINE Contract.
- Tempo contract and some messages : Work in progress.

+++ trunk/sources/shared/shared/plugin/yPluginApi/historization/Energy.cpp
@@ -6,7 +6,7 @@ namespace shared { namespace plugin { namespace yPluginApi { namespace historiza
 {
 
    CEnergy::CEnergy(const std::string& keywordName, const EKeywordAccessMode& accessMode, const EMeasureType& measureType, typeInfo::CDoubleTypeInfo & additionalInfo)
-      :CSingleHistorizableData<double>(keywordName, CStandardCapacities::Energy, accessMode, 0.0, measureType, additionalInfo)
+      :CSingleHistorizableData<unsigned long>(keywordName, CStandardCapacities::Energy, accessMode, 0, measureType, additionalInfo)
    {
    }
